add more checks
Added more checks for inaccurate dates. for ranges returns the average rounded down, otherwise the minimum.
For inaccurate dates adds the non-enumerable `.originalDate` property to the result.
For dates up to the second millennium tricks spacetime to get the right result.
If parsing was unsuccessful (`{ year: 1970, month: 0, date: 1 }`), returns null.