feat(image plugin): add `.license()` and redefine `.images()`. fix #451
I'm using `isomorphic-unfetch` and the main library's `_fetch` as the main source for HTTP requests.
made the new `api` folder for all the API related code in the image plugin.
changed `img-exists` to use unfetch as well and added userAgent info to the request.
redefined `images` which now also accepts an options object with a `batch` property. specifying method names with this property means the data for those methods (for all the images) is fetched once beforehand to prevent unnecessary calls. for the `exists` method the call is made to `iiprop = url` and the existence of a url property is checked.
the new `fetch.js` and `parse.js` are intended to serve as the central place for all the API related methods inside plugin.